+++ backend/src/main/java/com/company/jobmonitor/repository/UserRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByUsername(String username);
 
@@ -21,9 +21,9 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByEmail(String email);
 
-    boolean existsByUsernameAndIdNot(String username, Long id);
+    boolean existsByUsernameAndIdNot(String username, Integer id);
 
-    boolean existsByEmailAndIdNot(String email, Long id);
+    boolean existsByEmailAndIdNot(String email, Integer id);
 
     @Query("SELECT u FROM User u WHERE " +
            "(:search IS NULL OR :search = '' OR " +
